// packages/frontend/src/types/schemas/prisma-generated/inputTypeSchemas/NestedEnumPlanTypeNullableFilterSchema.ts
import type { Prisma } from '@prisma/client';

import { z } from 'zod';
import { PlanTypeSchema } from './PlanTypeSchema';

export const NestedEnumPlanTypeNullableFilterSchema: z.ZodType<Prisma.NestedEnumPlanTypeNullableFilter> = z.object({
  equals: z.lazy(() => PlanTypeSchema).optional().nullable(),
  in: z.lazy(() => PlanTypeSchema).array().optional().nullable(),
  notIn: z.lazy(() => PlanTypeSchema).array().optional().nullable(),
  not: z.union([ z.lazy(() => PlanTypeSchema),z.lazy(() => NestedEnumPlanTypeNullableFilterSchema) ]).optional().nullable(),
}).strict();

export default NestedEnumPlanTypeNullableFilterSchema;